# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AccountResponse" : {
        "properties" : {
          "balance" : {
            "type" : "string"
          },
          "iban" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "FROZEN" ],
            "type" : "string"
          },
          "type" : {
            "enum" : [ "CHECKING", "SAVINGS", "LOAN" ],
            "type" : "string"
          }
        },
        "required" : [ "balance", "iban", "id", "status", "type" ],
        "type" : "object"
      },
...
      "DepositRequest" : {
        "properties" : {
          "amount" : {
            "pattern" : "^\\d+(\\.\\d{1,4})?$",
            "type" : "string"
          }
        },
        "required" : [ "amount" ],
        "type" : "object"
      },
...
      "OpenAccountRequest" : {
        "properties" : {
          "type" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "type" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Full-stack online banking demo. Errors are RFC-7807 (ApiProblem); money travels as JSON strings.",
    "title" : "Online Banking Simulator API",
    "version" : "1.3.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/accounts" : {
      "get" : {
        "operationId" : "mine_1",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AccountResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "account-controller" ]
      },
      "post" : {
        "operationId" : "open",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OpenAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "account-controller" ]
      }
    },
    "/api/v1/accounts/{id}" : {
      "get" : {
        "operationId" : "detail",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "account-controller" ]
      }
    },
...
    "/api/v1/accounts/{id}/deposit" : {
      "post" : {
        "operationId" : "deposit",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DepositRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "account-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ]
}